mmt-47 Code style. Remove adding option to Helper/Config. Create Block.
