+++ app/code/Codifi/Sales/Block/System/Config/OrderTypeConfig.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * Codifi_Sales
+ *
+ * @copyright   Copyright (c) 2021 Codifi
+ * @author      Pavel Zelenevich <pzelenevich@codifi.me>
+ */
+
+namespace Codifi\Sales\Block\System\Config;
+
+use Magento\Config\Block\System\Config\Form\Field\FieldArray\AbstractFieldArray;
+
+/**
+ * Class OrderTypeConfig
+ * @package Codifi\Sales\Block\System\Config
+ */
+class OrderTypeConfig extends AbstractFieldArray
+{
+    /**
+     * Prepare rendering the new field by adding all the needed columns
+     */
+    protected function _prepareToRender()
+    {
+        $this->addColumn('value', ['label' => __('value'), 'class' => 'required-entry']);
+        $this->addColumn('description', ['label' => __('description'), 'class' => 'required-entry']);
+        $this->_addAfter = false;
+        $this->_addButtonLabel = __('Add');
+    }
+}
