+++ app/code/Codifi/Sales/Helper/Config.php
@@ -8,8 +8,8 @@
 
 namespace Codifi\Sales\Helper;
 
-use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\Store\Model\ScopeInterface;
+use Codifi\Sales\Setup\Patch\Data\AddOrderTypeToCoreConfigData;
+use Magento\Framework\Serialize\SerializerInterface;
 
 /**
  * Class Config
@@ -21,4 +21,42 @@ class Config
      * Order type attribute code
      */
     const ORDER_TYPE_CODE = 'order_type';
+
+    /**
+     * @var \Magento\Framework\App\Config
+     */
+    private $config;
+
+    /**
+     * @var SerializerInterface
+     */
+    private $serializer;
+
+    /**
+     * Config constructor.
+     *
+     * @param Config $config
+     * @param SerializerInterface $serializer
+     */
+    public function __construct(
+        Config $config,
+        SerializerInterface $serializer
+    )
+    {
+        $this->config = $config;
+        $this->serializer = $serializer;
+    }
+
+    /**
+     * Get all order_type options
+     *
+     * @return array
+     */
+    public function getAllOptions(): array
+    {
+        $value = $this->config->getValue(AddOrderTypeToCoreConfigData::XML_PATH_ORDER_TYPE_VALUE);
+        $options = $this->serializer->unserialize($value);
+
+        return $options;
+    }
 }
